[FE] Refactor : 파일구조 정리
- hook과 util 용도에 맞게 분리
- data 파일 삭제 -> constants에 병합
- 사용하지 않는 파일 삭제
