+++ frontend/src/components/Layout/PlayBar/index.tsx
@@ -13,7 +13,7 @@ import {
   IoCloseOutline,
 } from 'react-icons/io5';
 import { BsFillVolumeUpFill, BsMusicNoteList } from 'react-icons/bs';
-import useEventHandler from '@hooks/useEventHandler';
+import useEventHandler from '@utils/logEventHandler';
 import ClickEventWrapper from '@components/EventWrapper/ClickEventWrapper';
 
 import PlayTrackItem from '@components/Common/PlayTrackItem';
@@ -203,11 +203,22 @@ function PlayBar() {
 {JSON.stringify(log.eventTime)}
                     </Info>
                     <Info style={{ color: '#ffe500' }}>
-                      eventName : {JSON.stringify(log.eventName)}
+                      eventName : 
+{' '}
+{JSON.stringify(log.eventName)}
                     </Info>
-                    <Info>parameters :{JSON.stringify(log.parameters)}</Info>
-                    <Info>userInfo :{JSON.stringify(log.userInfo)}</Info>
-                    <Info>userAgent :{JSON.stringify(log.userAgent)}</Info>
+                    <Info>
+                      parameters :
+{JSON.stringify(log.parameters)}
+                    </Info>
+                    <Info>
+                      userInfo :
+{JSON.stringify(log.userInfo)}
+                    </Info>
+                    <Info>
+                      userAgent :
+{JSON.stringify(log.userAgent)}
+                    </Info>
                     <hr style={{ borderColor: 'darkgrey' }} />
                   </Log>
                 ))}
