+++ frontend/src/constants/description.ts
