+++ frontend/src/constants/lyricSample.ts
